# backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.ChangePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
  handlers.EditUserRequest:
    properties:
      is_super:
        type: boolean
      username:
        type: string
    type: object
  handlers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.Book:
    properties:
      author:
        type: string
      completed:
        type: boolean
      created_at:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Health:
    properties:
      status:
        type: string
      timestamp:
        type: string
    type: object
  models.User:
    properties:
      id:
        type: integer
      is_super:
        type: boolean
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing your reading list
  title: Want to Read API
  version: "1.0"
paths:
  /api/change_password:
    post:
      description: Change user password
      parameters:
      - description: Change password request
        in: body
        name: change_password
        required: true
        schema:
          $ref: '#/definitions/handlers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Change password
      tags:
      - users
  /api/login:
    post:
      description: Authenticate user and return JWT token
      parameters:
      - description: Login request
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Authenticate user
      tags:
      - users
  /api/register:
    post:
      description: Register a new user
      parameters:
      - description: Register request
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register user
      tags:
      - users
  /api/user/{id}:
    patch:
      description: Edit user details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Edit user request
        in: body
        name: edit_user
        required: true
        schema:
          $ref: '#/definitions/handlers.EditUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Edit user
      tags:
      - users
  /api/users:
    get:
      description: List all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List users
      tags:
      - users
  /api/users/{user_id}/book:
    post:
      description: Add a book into the database
      parameters:
      - description: Book to add
        in: body
        name: book
        required: true
        schema:
          type: object
      - description: ID of user
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Add book
      tags:
      - books
  /api/users/{user_id}/book/{id}:
    delete:
      description: Deletes a book based on the given ID
      parameters:
      - description: ID of book
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete book
      tags:
      - books
    patch:
      description: Updates a book based on the given ID
      parameters:
      - description: ID of book
        in: path
        name: id
        required: true
        type: integer
      - description: Adjusted book object
        in: body
        name: book
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update book
      tags:
      - books
  /api/users/{user_id}/books/{id}:
    get:
      description: Get a list of all books in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Book'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List all books
      tags:
      - books
  /health:
    get:
      description: Get the health status of the API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Health'
      summary: Health check endpoint
      tags:
      - health
swagger: "2.0"